requirements.txt
Pillow
requests
mysql-connector-python
eel